# Optimized IsZero and IsOne